style(templates): update padding and class names for consistency
Set `padding_bottom` to 0 across multiple JSON templates to ensure
consistent spacing and layout. This change helps in maintaining a
uniform design across different sections of the application.

Add `page-content` and `page-content--flush-footer` classes to
various Liquid templates to standardize the styling and improve
the maintainability of the codebase. These class additions help
in achieving a consistent look and feel across different pages.